some cleanup, updated softmath with some google results :), updated readme

+++ libsrc/float/softmath/softmath-truncf.c
@@ -1,7 +1,8 @@
 
 #include <math.h>
 
+/* FIXME: this is really too simple */
 float truncf(float x)
 {
-    return (float)((signed)x);
+    return (float)((signed long)x);
 }
